requirements.txt
streamlit
pandas
plotly
requests
streamlit-autorefresh
fastapi
uvicorn
pydantic
httpx
python-dotenv
google-generativeai
fastmcp
